+++ src/Tickets/Commerce/Provider.php
@@ -68,6 +68,8 @@ class Provider extends tad_DI52_ServiceProvider {
 		$this->container->singleton( Checkout::class );
 		$this->container->singleton( Settings::class );
 		$this->container->singleton( Tickets_View::class );
+		$this->container->singleton( Promoter_Observer::class );
+		tribe( Promoter_Observer::class );
 
 		$this->container->register( Status\Status_Handler::class );
 		$this->container->register( Flag_Actions\Flag_Action_Handler::class );
